Remove useless static_cast
git-svn-id: svn://scribus.net/branches/Version15x/Scribus@25488 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scpageoutput.cpp
@@ -107,7 +107,7 @@ void ScPageOutput::drawMasterItems(ScPainterExBase *painter, ScPage *page, ScLay
 		currItem = page->FromMaster.at(i);
 		if (currItem->m_layerID != layer.ID)
 			continue;
-		if ((currItem->OwnPage != -1) && (currItem->OwnPage != static_cast<int>(Mp->pageNr())))
+		if ((currItem->OwnPage != -1) && (currItem->OwnPage != Mp->pageNr()))
 			continue;
 		if (!currItem->printEnabled())
 			continue;
